added updates from master

+++ libraries/my_bicycle_lib/src/buzzer.h
@@ -1,4 +1,4 @@
 
 void init_buzzer();
 
-void enable_buzzer(unsigned int ms_sound, unsigned int repeat = 1, unsigned int delay_ms = 0);
+void enable_buzzer(unsigned long ms_sound, unsigned int repeat = 1, unsigned int delay_ms = 0);
